requirements.txt
streamlit
os
sys
pillow